Add files via upload
Creates a custom control for the Customizer in WordPress. The control creates an item with multiple inputs for image, title, description and page link. User can add, edit or delete items in Customizer. Created with intent of displaying services for a business, and giving user flexibility in how many services they wanted to display on page without needing to know how to code.